ci replace mysqladmin
Recent container releases deprecated mysqladmin and use mariadb-admin.
The server code will eventually follow suite.

In this change, we use the container healthcheck.sh script to
use the mysql@localhost user and check to see if the connection
is there on TCP (which it didn't before), and that the
install-wp-tests.sh will use mariadb-admin and fall back to mysqladmin
if needed.

The MYSQL_ROOT_PASSWORD = '' was implied by the
{MYSQL|MARIADB}_ALLOW_EMPTY{_ROOT}_PASSWORD.

Signed-off-by: Daniel Black <daniel@mariadb.org>